@@ -70,9 +70,9 @@
     "mongodb": "^3.1.4",
     "nodemailer": "^6.7.3",
     "query-string": "^6.1.0",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-chartjs-2": "^2.7.4",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0",
     "react-dropzone": "^5.0.1",
     "react-helmet": "^6.1.0",
     "react-lazyload": "^2.6.5",
